backend/requirements.txt
asgiref
Django == 3.1.7
django-cors-headers
djangorestframework
psycopg2
pytz
sqlparse
redis
channels == 3.0.3
channels-redis==3.2.0
pika
